default.runner/prefect-ef/requirements.txt
atlassian-python-api
awswrangler
boto3
beautifulsoup4
dbt-snowflake
geopy
googlemaps
openpyxl
oracledb
# We need a newer version of oscrypto than the one released to get a fix for a bug
# that gets triggered when having double digits on the OpenSSL version
# (https://github.com/wbond/oscrypto/issues/75).
# We can remove the oscrypto pinning once the fix becomes part of a new release.
oscrypto@git+https://github.com/wbond/oscrypto.git@d5f3437ed24257895ae1edd9e503cfb352e635a8
paramiko==3.4.0
phpserialize
Pillow
prefect-aws
prefect-dbt==0.4.3
prefect-fivetran
prefect-github
prefect-shell
prefect-email
prefect==2.*
psutil
pydantic==1.10.13
pyjwt
pymongo==3.13
pymongo-auth-aws
pymssql
pyodbc
pysftp
ratelimit
requests
s3fs
shareplum
snowflake-connector-python[pandas]==3.0.3
snowflake-sqlalchemy==1.4.7
statsmodels
xmltodict
aiohttp
aioboto3
numpy==1.26.4
reportlab
polars
